requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
imblearn
sqlalchemy
psycopg2-binary
python-dotenv
jupyter